# app/static/swagger.yaml
...
swagger: "2.0"
...
info:
  title: "Mechanic API" # Title of the API
  description: "API to manage a Mechanic Shop" # A brief description of what the API does
  version: "1.0.0" # The version of your API

# The base URL of the API, in this case, "127.0.0.1:5000". This is where the API can be accessed.
host: "127.0.0.1:5000"
...
schemes:
  - "http"

# The types of request body formats the API can consume. Here, it's "application/json",
# meaning the API can accept JSON-formatted request bodies.
consumes:
  - "application/json"

# The types of response body formats the API can produce. Here, it's also "application/json",
# indicating that the API will respond with JSON-formatted data.
produces:
  - "application/json"

# Add securityDefinitions to allow endpoints to require jwt (i.e. tokens) for access
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# Paths - This is where you would define the endpoints of your API, along with the HTTP methods
# they support and their respective request/response formats.

# -------------------- Customer Endpoints --------------------
paths:
...
  /service_tickets:
    post:
      tags:
        - Service Tickets
      summary: "Endpoint to create/register a new service ticket"
      description: "Endpoint to create a new service ticket in our Mechanic Shop system."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new service ticket"
          required: true
          schema:
            $ref: "#/definitions/CreateTicketPayload"
      responses:
        201:
          description: "Service ticket created successfully"
          schema:
            $ref: "#/definitions/CreateTicketResponse"
          examples:
            application/json:
              id: 1
              VIN: "0123455890"
              service_date: "2025-12-10"
              service_desc: "Tire Repair"
              customer_id: 1
              mechanic_ids: [1, 2]

    get: # ⚡ Tested!
      tags:
        - Service Tickets
      summary: "Returns all service tickets"
      description: "Endpoint to retrieve a list of service tickets in the Mechanic Shop system."
      responses:
        200:
          description: "Retrieved service tickets successfully"
          schema:
            $ref: "#/definitions/AllTickets"
  /service_tickets/{service_ticket_id}: # ⚡ Tested!
    get:
      tags:
        - Service Tickets
      summary: "Get a specific service ticket by ID"
      description: "Retrieve a specific service ticket using the unique ID."
      parameters:
        - name: service_ticket_id
          in: path
          required: true
          description: "ID of the service ticket to retrieve"
          type: integer
      responses:
        200:
          description: "Service ticket retrieved successfully"
          schema:
            $ref: "#/definitions/CreateTicketResponse"
          examples:
            application/json:
              id: 1
              VIN: "0123455890"
              service_date: "2025-12-10"
              service_desc: "Tire Repair"
              customer_id: 1
              mechanic_ids: [1, 2]
        404:
          description: "Service ticket not found"

    put: # ⚡ Tested!
      tags:
        - Service Tickets
      summary: "Update a service ticket (by ID) by adding or removing mechanic(s)"
      description: "Retrieve a specific service ticket using the unique ID and then add or remove mechanics. Returns a message and full service ticket info."
      parameters:
        - name: service_ticket_id
          in: path
          required: true
          description: "ID of the service ticket to update"
          type: integer
        - in: body
          name: body
          description: "Mechanic IDs to add or remove"
          required: true
          schema:
            $ref: "#/definitions/UpdateTicketPayload"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            $ref: "#/definitions/UpdateTicketResponse"
          examples:
            application/json:
              message: "Mechanics [3,5] added/removed successfully."
              service_ticket:
                id: 1
                VIN: "0123455890"
                service_date: "2025-12-10"
                service_desc: "Tire Repair"
                customer_id: 1
                mechanic_ids: [1, 2, 3, 5]
        404:
          description: "Service ticket not found"
  /service_tickets/my-tickets: # ⚡ Tested!
    get:
      tags:
        - Service Tickets
      summary: "Get all service tickets for the logged in customer"
      description: "Retrieve all service tickets associated with the logged in customer. Token required"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Service tickets retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CreateTicketResponse"
  /service_tickets/{service_ticket_id}/add-part: # ⚡ Tested!
    put:
      tags:
        - Service Tickets
      summary: "Update a service ticket (by ID) by adding a part from inventory"
      description: "Retrieve a specific service ticket using the unique ID and then add a part from inventory."
      parameters:
        - name: service_ticket_id
          in: path
          required: true
          description: "ID of the service ticket to retrieve"
          type: integer
        - in: "body"
          name: "body"
          description: "Information required to update the service ticket"
          required: true
          schema:
            $ref: "#/definitions/UpdatePartPayload"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            $ref: "#/definitions/UpdatePartResponse"
          examples:
            application/json:
              message: "Part <part_name> added successfully."
              status: "success"
        404:
          description: "Service ticket not found"

# This is where you would define the data structures used in your API,
# such as request and response bodies.
definitions:
...
  CreateTicketPayload:
    type: "object"
    properties:
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      service_desc:
        type: "string"
      customer_id:
        type: "integer"
      mechanic_ids:
        type: "array"
        items:
          type: "integer"
    required:
      - VIN
      - service_date
      - service_desc
      - customer_id
      - mechanic_ids
  CreateTicketResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      service_desc:
        type: "string"
      customer_id:
        type: "integer"
      mechanic_ids:
        type: "array"
        items:
          type: "integer"
  AllTickets:
    type: "array"
    items:
      type: "object"
      id:
        type: "integer"
      VIN:
        type: "string"
      service_date:
        type: "string"
        format: "date"
      service_desc:
        type: "string"
      customer_id:
        type: "integer"
      mechanic_ids:
        type: "array"
        items:
          type: "integer"
  UpdateTicketPayload:
    type: "object"
    properties:
      add_ids:
        type: "integer"
      remove_ids:
        type: "integer"
    required:
      - add_ids
      - remove_ids
  UpdateTicketResponse:
    type: object
    properties:
      message:
        type: string
        description: "Information about which mechanic IDs were added or removed"
      service_ticket:
        type: object
        properties:
          id:
            type: integer
          VIN:
            type: string
          service_date:
            type: string
            format: date
          service_desc:
            type: string
          customer_id:
            type: integer
          mechanic_ids:
            type: array
            items:
              type: integer
    required:
      - message
      - service_ticket
  UpdatePartPayload:
    type: "object"
    properties:
      part_id:
        type: "integer"
    required:
      - part_id
  UpdatePartResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
    required:
      - message
      - status